requirements.txt
alembic==1.13.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
docutils==0.20.1
entrypoints==0.4
Flask==3.0.0
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mlflow==2.9.2
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pdf2image==1.17.0
pillow==10.2.0
pkgconfig==1.5.5
protobuf==4.25.2
psutil==5.9.7
pyarrow==14.0.2
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
tabulate==0.9.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
websocket-client==1.7.0
Werkzeug==3.0.1
zipp==3.17.0
